+++ src/cpp/rtps/writer/RTPSWriter.cpp
@@ -415,6 +415,8 @@ bool RTPSWriter::is_pool_initialized() const
     return true;
 }
 
+#ifdef FASTDDS_STATISTICS
+
 bool RTPSWriter::add_statistics_listener(
             std::shared_ptr<fastdds::statistics::IListener> listener)
 {
@@ -428,6 +430,8 @@ bool RTPSWriter::remove_statistics_listener(
      (void)listener;
      return false;
 }
+
+#endif // FASTDDS_STATISTICS
 
 }  // namespace rtps
 }  // namespace fastrtps
